+++ tests/cli/error_value.rs
@@ -99,8 +99,8 @@ fn warn_if_error_value_starts_with_err() {
             src/lib.rs:3: replace even_is_ok -> Result<u32, &\'static str> with Ok(0)
             src/lib.rs:3: replace even_is_ok -> Result<u32, &\'static str> with Ok(1)
             src/lib.rs:3: replace even_is_ok -> Result<u32, &\'static str> with Err(Err(anyhow!(\"mutant\")))
-        "
-        });
+            src/lib.rs:4: replace == with != in even_is_ok
+        " });
 }
 
 #[test]